% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{airbrake.notify.and.raise}
\alias{airbrake.notify.and.raise}
\title{Send an Airbrake notification on error and then raise the error again. Useful
for scripts where the proper OS exit codes are relevant.}
\usage{
airbrake.notify.and.raise(error)
}
\description{
Send an Airbrake notification on error and then raise the error again. Useful
for scripts where the proper OS exit codes are relevant.
}
