% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{airbrake.notify}
\alias{airbrake.notify}
\title{Send an Airbrake notification on error. Example `tryCatch(stop('error'), error=airbrake.notify)`}
\usage{
airbrake.notify(error)
}
\description{
Send an Airbrake notification on error. Example `tryCatch(stop('error'), error=airbrake.notify)`
}
